requirements.txt
streamlit==1.31.0
snowflake-snowpark-python
ConfigParser==5.3.0
PyPDF2==3.0.1
langchain==0.1.1
